requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
